Implement mark channels read endpoints (#35)
* Implement mark channels read state endpoint + MessageRead and NotificationMarkRead events

* Implement endpoint for marking single channel read state + add integration test

* Add missing EventMessageNew fields + Add sample project option to mark message as last read & logs for MessageRead & NotificationMarkRead events

+++ Assets/Plugins/StreamChat/Core/DTO/Models/ReadDTO.cs
@@ -20,7 +20,7 @@ namespace StreamChat.Core.DTO.Models
         public System.DateTimeOffset? LastRead { get; set; }
 
         [Newtonsoft.Json.JsonProperty("unread_messages", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? UnreadMessages { get; set; }
+        public int? UnreadMessages { get; set; }
 
         [Newtonsoft.Json.JsonProperty("user", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public UserObjectDTO User { get; set; }
